Tests/LibWeb/Text/expected/async-scrolling/hidden-axis-programmatic-scroll-preserved.txt

initial scrollLeft: 50
vertical wheel target: non-viewport
horizontal wheel target: none
final scrollLeft: 50
final scrollTop: 100
